requirements.txt
flask>=3.0
yt-dlp>=2024.0
requests>=2.31
beautifulsoup4>=4.12
pywebview>=5.0
curl_cffi>=0.7
